requirements.txt
requests>=2.20.0
flask>=2.0.0
PyYAML
rich
PyDDENT @ git+https://github.com/NIH-NCPI/PyDDENT
